refactor: move auth guard to shared utility
