+++ frontend/src/routes/containers/$containerId/logs.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { createFileRoute, redirect, useNavigate } from "@tanstack/react-router";
+import { createFileRoute, useNavigate } from "@tanstack/react-router";
 import {
   ArrowLeftIcon,
   CheckIcon,
@@ -55,39 +55,11 @@ import {
   getStateBadgeClass,
   toTitleCase
 } from "@/features/containers/components/container-utils";
-import { API_BASE_URL } from "@/types/api";
+import { requireAuthIfEnabled } from "@/lib/auth-guard";
 
 export const Route = createFileRoute("/containers/$containerId/logs")({
   beforeLoad: async () => {
-    const token = localStorage.getItem("logdeck_auth_token");
-
-    // If no token, check if auth is required
-    if (!token) {
-      try {
-        const authUrl = `${API_BASE_URL}/api/v1/auth/login`.replace(
-          /([^:]\/)\/+/g,
-          "$1"
-        );
-        const response = await fetch(authUrl, {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({ username: "", password: "" }),
-        });
-
-        // If 404, auth is disabled - allow access
-        if (response.status === 404) {
-          return;
-        }
-
-        // Auth is enabled but no token - redirect to login
-        throw redirect({ to: "/login" });
-      } catch (error) {
-        // If we can't reach the server, allow access (fail open for development)
-        if (error instanceof Error && error.message.includes("redirect")) {
-          throw error;
-        }
-      }
-    }
+    await requireAuthIfEnabled();
   },
   component: ContainerLogsPage,
 });
